// frontend/src/types/content.ts
export interface NavItem {
  label: string;
  href: string;
}

export interface FeatureItem {
  title: string;
  description: string;
  icon: string;
}

export interface StepItem {
  title: string;
  description: string;
  icon: string;
}

export interface CropItem {
  name: string;
  tag: string;
}

export interface StatItem {
  value: string;
  label: string;
}

export interface WeatherItem {
  label: string;
  value: string;
  symbol: string;
}